File uploading finished

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,23 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
-// use DB;
+use DB;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -45,14 +57,33 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body'  => 'required'
+            'body'  => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload
+        if ($request->hasFile('cover_image')) {
+            // Get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . "_" . time() . "." . $extension;
+            // Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create Posts
         $post = new Post;
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
         $post->user_id  = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created successfully');
@@ -79,6 +110,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -96,10 +132,28 @@ class PostsController extends Controller
             'body'  => 'required'
         ]);
 
+        // Handle file upload
+        if ($request->hasFile('cover_image')) {
+            // Get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . "_" . time() . "." . $extension;
+            // Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
+
         // Create Posts
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated successfully');
@@ -114,6 +168,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if ($post->cover_image != 'noimage.jpg') {
+            // Delete image
+            Storage::delete('public/cover_images' . $post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted');
     }
